refactor: rename components as per convention and use appropriate extensions
