+++ src/components/frameControllers/DetailsInput.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+
+// Components
+import Upload from '../toolbox/Upload';
+import CustomText from '../toolbox/InputText';
+
+function Inputs({
+  uploadedImage,
+  setUploadedImage,
+  userName,
+  setUsername,
+  guildName,
+  setGuildname,
+}) {
+  return (
+    <div>
+      <Upload uploadedImage={uploadedImage} setUploadedImage={setUploadedImage} />
+      <CustomText
+        userName={userName}
+        setUsername={setUsername}
+        guildName={guildName}
+        setGuildname={setGuildname}
+      />
+    </div>
+  );
+}
+
+export default Inputs;
